kernel_pairwise_sum: Fix issue with even->odd vector length
When the summation went from an even number of elements in the starting
vector to an odd number after the first layer of operations, the last
element was silently discarded. This makes the summation more robust and
now can handle an odd or even number of elements in each layer.